del: unused files add: split settings fix: docker, env
