--- .envs/.local/.django
@@ -1,7 +1,0 @@
-CELERY_BROKER=redis://redis:6379/0
-
-DOMAIN=localhost:8080
-EMAIL_PORT=1025
-CELERY_FLOWER_USER=admin
-CELERY_FLOWER_PASSWORD=123qwe
-SIGNING_KEY=UH89LJUWuo4xBA4E9fEe1Vz11Hch3hp5aBthPGsn7N5dBWpcpn0
