+++ .gitignore
@@ -159,3 +159,5 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
 !/.envs/.production/
+/mediafiles/
+/staticfiles/
