+++ authors_api/celery.py
@@ -2,8 +2,7 @@ import os
 from celery import Celery
 from django.conf import settings
 
-# TODO: change it  production
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'authors_api.settings.local')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'authors_api.settings')
 
 app = Celery('authors_api')
