+++ authors_api/components/celery_email.py
@@ -1,0 +1,8 @@
+from authors_api.settings import env
+
+EMAIL_BACKEND = "djcelery_email.backends.CeleryEmailBackend"
+EMAIL_HOST = env("EMAIL_HOST", default="mailhog")
+EMAIL_PORT = env("EMAIL_PORT")
+DEFAULT_FROM_EMAIL = "support@apiimperfect.site"
+DOMAIN = env("DOMAIN")
+SITE_NAME = "Authors Haven"
