+++ authors_api/components/database.py
@@ -1,0 +1,3 @@
+from authors_api.settings import env
+
+DATABASES = {"default": env.db("DATABASE_URL")}
