+++ authors_api/components/elasticsearch.py
@@ -1,0 +1,5 @@
+ELASTICSEARCH_DSL = {
+    "default": {
+        "hosts": "elasticsearch:9200",
+    },
+}
