+++ authors_api/components/internationalization.py
@@ -1,0 +1,7 @@
+LANGUAGE_CODE = "en-us"
+
+TIME_ZONE = "UTC"
+
+USE_I18N = True
+
+USE_TZ = True
