+++ authors_api/settings.py
@@ -1,0 +1,73 @@
+"""
+Django settings for authors_api project.
+
+Generated by 'django-admin startproject' using Django 4.2.5.
+
+For more information on this file, see
+https://docs.djangoproject.com/en/4.2/topics/settings/
+
+For the full list of settings and their values, see
+https://docs.djangoproject.com/en/4.2/ref/settings/
+"""
+
+from pathlib import Path
+
+import environ
+from split_settings.tools import include
+
+env = environ.Env()
+
+ROOT_DIR = Path(__file__).resolve().parent.parent.parent
+
+APP_DIR = ROOT_DIR / "core_apps"
+
+DEBUG = env.bool("DEBUG", False)
+
+SECRET_KEY = env(
+    "DJANGO_SECRET_KEY",
+    default="BRYV5OIQAXkEfWrTMpMgTJ2tk6Z5_pvUBtAGlb0eq98pSitCo7w",
+)
+
+CSRF_TRUSTED_ORIGINS = ["http://localhost:8080"]
+
+include(
+    "components/allauth.py",
+    "components/auth_password_components.py",
+    "components/celery.py",
+    "components/celery_email.py",
+    "components/database.py",
+    "components/elasticsearch.py",
+    "components/installed_apps.py",
+    "components/internationalization.py",
+    "components/jwt.py",
+    "components/logging.py",
+    "components/middlewares.py",
+    "components/rest_framework.py",
+    "components/templates.py",
+)
+
+ROOT_URLCONF = "authors_api.urls"
+
+WSGI_APPLICATION = "authors_api.wsgi.application"
+
+SITE_ID = 1
+
+ADMIN_URL = "supersecret/"
+
+# Static files (CSS, JavaScript, Images)
+# https://docs.djangoproject.com/en/4.2/howto/static-files/
+
+STATIC_URL = "/staticfiles/"
+STATIC_ROOT = str(ROOT_DIR / "staticfiles")
+
+MEDIA_URL = "/mediafiles/"
+MEDIA_ROOT = str(ROOT_DIR / "mediafiles")
+
+# Default primary key field type
+# https://docs.djangoproject.com/en/4.2/ref/settings/#default-auto-field
+
+DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
+
+CORS_URLS_REGEX = r"^/api/.*$"
+
+AUTH_USER_MODEL = "users.User"
