--- authors_api/settings/base.py
@@ -1,242 +1,0 @@
-"""
-Django settings for authors_api project.
-
-Generated by 'django-admin startproject' using Django 4.2.5.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.2/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/4.2/ref/settings/
-"""
-
-from pathlib import Path
-from datetime import timedelta
-import environ
-
-env = environ.Env()
-
-# Build paths inside the project like this: BASE_DIR / 'subdir'.
-ROOT_DIR = Path(__file__).resolve().parent.parent.parent
-
-APP_DIR = ROOT_DIR / "core_apps"
-
-DEBUG = env.bool("DJANGO_DEBUG", False)
-# Application definition
-
-DJANGO_APPS = [
-    "django.contrib.admin",
-    "django.contrib.auth",
-    "django.contrib.contenttypes",
-    "django.contrib.sessions",
-    "django.contrib.messages",
-    "django.contrib.staticfiles",
-    "django.contrib.sites",
-]
-
-THIRD_PARTY_APPS = [
-    "rest_framework",
-    "django_filters",
-    "django_countries",
-    "phonenumber_field",
-    "drf_yasg",
-    "corsheaders",
-    "djcelery_email",
-    "rest_framework.authtoken",
-    "allauth",
-    "allauth.account",
-    "allauth.socialaccount",
-    "dj_rest_auth",
-    "dj_rest_auth.registration",
-    "taggit",
-    "django_elasticsearch_dsl",
-    "django_elasticsearch_dsl_drf",
-]
-
-LOCAL_APPS = [
-    "core_apps.profiles",
-    "core_apps.common",
-    "core_apps.users",
-    "core_apps.articles",
-    "core_apps.ratings",
-    "core_apps.bookmarks",
-    "core_apps.responses",
-    "core_apps.search",
-]
-
-INSTALLED_APPS = [
-    *DJANGO_APPS,
-    *THIRD_PARTY_APPS,
-    *LOCAL_APPS,
-]
-
-MIDDLEWARE = [
-    "django.middleware.security.SecurityMiddleware",
-    "corsheaders.middleware.CorsMiddleware",
-    "django.contrib.sessions.middleware.SessionMiddleware",
-    "django.middleware.common.CommonMiddleware",
-    "django.middleware.csrf.CsrfViewMiddleware",
-    "django.contrib.auth.middleware.AuthenticationMiddleware",
-    "django.contrib.messages.middleware.MessageMiddleware",
-    "django.middleware.clickjacking.XFrameOptionsMiddleware",
-    "allauth.account.middleware.AccountMiddleware",
-]
-
-ROOT_URLCONF = "authors_api.urls"
-
-TEMPLATES = [
-    {
-        "BACKEND": "django.template.backends.django.DjangoTemplates",
-        "DIRS": [],
-        "APP_DIRS": True,
-        "OPTIONS": {
-            "context_processors": [
-                "django.template.context_processors.debug",
-                "django.template.context_processors.request",
-                "django.contrib.auth.context_processors.auth",
-                "django.contrib.messages.context_processors.messages",
-            ],
-        },
-    },
-]
-
-WSGI_APPLICATION = "authors_api.wsgi.application"
-
-DATABASES = {"default": env.db("DATABASE_URL")}
-
-PASSWORD_HASHERS = [
-    "django.contrib.auth.hashers.Argon2PasswordHasher",
-    "django.contrib.auth.hashers.PBKDF2PasswordHasher",
-    "django.contrib.auth.hashers.PBKDF2SHA1PasswordHasher",
-    "django.contrib.auth.hashers.BCryptSHA256PasswordHasher",
-    "django.contrib.auth.hashers.ScryptPasswordHasher",
-]
-# Password validation
-# https://docs.djangoproject.com/en/4.2/ref/settings/#auth-password-validators
-
-AUTH_PASSWORD_VALIDATORS = [
-    {
-        "NAME": "django.contrib.auth.password_validation.UserAttributeSimilarityValidator",
-    },
-    {
-        "NAME": "django.contrib.auth.password_validation.MinimumLengthValidator",
-    },
-    {
-        "NAME": "django.contrib.auth.password_validation.CommonPasswordValidator",
-    },
-    {
-        "NAME": "django.contrib.auth.password_validation.NumericPasswordValidator",
-    },
-]
-
-# Internationalization
-# https://docs.djangoproject.com/en/4.2/topics/i18n/
-
-LANGUAGE_CODE = "en-us"
-
-TIME_ZONE = "UTC"
-
-USE_I18N = True
-
-USE_TZ = True
-
-SITE_ID = 1
-
-ADMIN_URL = "supersecret/"
-
-# Static files (CSS, JavaScript, Images)
-# https://docs.djangoproject.com/en/4.2/howto/static-files/
-
-STATIC_URL = "/staticfiles/"
-STATIC_ROOT = str(ROOT_DIR / "staticfiles")
-
-MEDIA_URL = "/mediafiles/"
-MEDIA_ROOT = str(ROOT_DIR / "mediafiles")
-
-# Default primary key field type
-# https://docs.djangoproject.com/en/4.2/ref/settings/#default-auto-field
-
-DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
-
-CORS_URLS_REGEX = r"^/api/.*$"
-
-ELASTICSEARCH_DSL = {
-    "default": {
-        "hosts": "es:9200",
-    },
-}
-
-LOGGING = {
-    "version": 1,
-    "disable_existing_loggers": False,
-    "formatters": {
-        "verbose": {
-            "format": "%(levelname)s %(name)-12s %(asctime)s %(module)s "
-                      "%(process)d %(thread)d %(message)s"
-        }
-    },
-    "handlers": {
-        "console": {
-            "level": "DEBUG",
-            "class": "logging.StreamHandler",
-            "formatter": "verbose",
-        }
-    },
-    "root": {"level": "INFO", "handlers": ["console"]},
-}
-
-AUTH_USER_MODEL = "users.User"
-
-CELERY_BROKER_URL = env("CELERY_BROKER")
-CELERY_RESULT_BACKEND = env("CELERY_BROKER")
-CELERY_ACCEPT_CONTENT = ["json"]
-CELERY_TASK_SERIALIZER = "json"
-CELERY_RESULT_SERIALIZER = "json"
-CELERY_RESULT_BACKEND_MAX_RETRIES = 10
-CELERY_TASK_SEND_SENT_EVENT = True
-
-if USE_TZ:
-    CELERY_TIMEZONE = TIME_ZONE
-
-REST_FRAMEWORK = {
-    "DEFAULT_AUTHENTICATION_CLASSES": [
-        "dj_rest_auth.jwt_auth.JWTCookieAuthentication",
-    ],
-    "DEFAULT_PERMISSION_CLASSES": [
-        "rest_framework.permissions.IsAuthenticated",
-    ],
-    "DEFAULT_FILTER_BACKENDS": [
-        "django_filters.rest_framework.DjangoFilterBackend",
-    ]
-}
-
-SIMPLE_JWT = {
-    "AUTH_HEADER_TYPES": ("Bearer",),  # Ожидаемый тип заголовка для JWT токена.
-    "ACCESS_TOKEN_LIFETIME": timedelta(minutes=30),
-    "REFRESH_TOKEN_LIFETIME": timedelta(days=1),
-    "ROTATE_REFRESH_TOKENS": True,
-    "SIGNING_KEY": env("SIGNING_KEY"),  # Секретный ключ для подписи JWT токена.
-    "USER_ID_FIELD": "id",  # Поле модели пользователя, которое будет использоваться для идентификации пользователя.
-    "USER_ID_CLAIM": "user_id",  # Ключ в токене, который будет использоваться для идентификации пользователя.
-}
-
-REST_AUTH = {
-    "USE_JWT": True,  # Использовать JWT токен для аутентификации.
-    "JWT_AUTH_COOKIE": "authors-access-token",
-    "JWT_AUTH_REFRESH_COOKIE": "authors-refresh-token",
-    "REGISTER_SERIALIZER": "core_apps.users.serializers.CustomRegisterSerializer",
-    # Использовать кастомный сериализатор для регистрации.
-}
-
-AUTHENTICATION_BACKENDS = [
-    "allauth.account.auth_backends.AuthenticationBackend",
-    "django.contrib.auth.backends.ModelBackend",
-]
-
-ACCOUNT_AUTHENTICATION_METHOD = "email"  # Использовать email для аутентификации.
-ACCOUNT_EMAIL_REQUIRED = True  # Обязательно указывать email при регистрации.
-ACCOUNT_EMAIL_VERIFICATION = "mandatory"  # Обязательно подтверждать email.
-ACCOUNT_CONFIRM_EMAIL_ON_GET = True  # Подтверждать email при переходе по ссылке.
-ACCOUNT_EMAIL_CONFIRMATION_EXPIRE_DAYS = 1  # Срок действия ссылки для подтверждения email.
-ACCOUNT_USER_MODEL_USERNAME_FIELD = None  # Не использовать username.
-ACCOUNT_USERNAME_REQUIRED = False  # Не использовать username.
