--- authors_api/settings/local.py
@@ -1,19 +1,0 @@
-from .base import *  # noqa
-from .base import env
-
-SECRET_KEY = env(
-    "DJANGO_SECRET_KEY",
-    default="BRYV5OIQAXkEfWrTMpMgTJ2tk6Z5_pvUBtAGlb0eq98pSitCo7w",
-)
-
-# SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = True
-
-CSRF_TRUSTED_ORIGINS = ["http://localhost:8080"]
-
-EMAIL_BACKEND = "djcelery_email.backends.CeleryEmailBackend"
-EMAIL_HOST = env("EMAIL_HOST", default="mailhog")
-EMAIL_PORT = env("EMAIL_PORT")
-DEFAULT_FROM_EMAIL = "support@apiimperfect.site"
-DOMAIN = env("DOMAIN")
-SITE_NAME = "Authors Haven"
