--- authors_api/settings/production.py
@@ -1,6 +1,0 @@
-from .base import *  # noqa
-from .base import env
-
-ADMIN = [("Sergei Aleshin", "test@test.com")]
-
-CSRF_TRUSTED_ORIGINS = [""]
