+++ authors_api/wsgi.py
@@ -11,7 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-# TODO: change this in production
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "authors_api.settings.local")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "authors_api.settings")
 
 application = get_wsgi_application()
