+++ docker/django/celery/worker/start.sh
