+++ docker/nginx/Dockerfile
@@ -2,4 +2,4 @@ FROM nginx:1.23.3-alpine
 
 RUN rm /etc/nginx/conf.d/default.conf
 
-COPY ./default.conf /etc/nginx/conf.d/default.conf
+COPY default.conf /etc/nginx/conf.d/default.conf
