+++ requirements.txt
@@ -21,3 +21,6 @@ djangorestframework-simplejwt==5.3.0
 dj-rest-auth==5.0.1
 django-elasticsearch-dsl==7.3.0
 django-elasticsearch-dsl-drf==0.22.5
+split_settings==1.0.0
+psycopg2-binary==2.9.9
+watchfiles==0.21.0
