--- requirements/local.txt
@@ -1,8 +1,0 @@
--r base.txt
-
-psycopg2-binary==2.9.9
-
-flake8==6.1.0
-black==23.9.1
-isort==5.12.0
-watchfiles==0.21.0
